Implements setExpression request.
setExpression is emitted by Visual Studio to change a value of a watch variable. Such request is out of the definition of VSCode debug protocol.

In this PR we implement only for Chrome. It is what we have verified working. Other debug adapter might need a different approach to update the value.