Add workaround for race condition
We discovered a race condition that doesn't happen often when testing locally, but seems to happen quite often on the CI machines. The issue occurs if a breakpoint set request comes in after the first scriptParsed event, but before the target script is parsed, the adapter tries to use breakOnLoad with an instrumentation breakpoint, but because scripts are already running on the browser, that instrumentation breakpoint will never be hit, and thus the pending breakpoints will never be resolved.

Note that this commit is only a workaround for the tests, and does not fix the underlying issue, which is that breakpoints set during a specific window of time will never actually be set in Chrome.